+++ client/src/app.css
@@ -5,6 +5,4 @@
   align-items: center;
   justify-content: space-between;
   min-height: 100vh;
-  width: 80%;
-  /* gap: 1rem; */
 }
