+++ client/src/components/Footer.module.css
@@ -1,4 +1,5 @@
 .footer_container {
+  margin-top: 1rem;
   background-color: #5825cc;
   width: 100%;
   height: 100px;
