+++ client/src/components/NavBar.module.css
@@ -1,9 +1,11 @@
 .navBar {
-  width: 100%;
-  height: 15vh;
+  padding: 10px;
+  width: 80%;
+  height: 10vh;
   display: flex;
   justify-content: space-between;
   align-items: center;
+  /* margin-bottom: 1rem; */
 }
 
 .navBar a:first-child {
