+++ client/src/views/Home.module.css
@@ -4,7 +4,7 @@
   align-items: center;
   justify-content: center;
   gap: 2rem;
-  width: 100%;
+  width: 80%;
 }
 
 .filtros_container {
